Move stockage and scan related components outside /containers
