+++ components/projet/livrables-section.js
@@ -2,7 +2,7 @@ import {useState} from 'react'
 import PropTypes from 'prop-types'
 import {sortBy} from 'lodash-es'
 
-import StockagePreview from '../containers/stockage-preview.js'
+import StockagePreview from './stockage-preview.js'
 import {LIVRABLE_NATURES} from '@/lib/utils/projet.js'
 
 import {livrableRenderItem} from '@/components/projet/list-render-items.js'
